Factor cursor correction logic into own files
Trying to cut down on the size of the BlockEditorEditing component.
